requirements.txt
numpy==2.3.1
pillow==11.3.0
pygame==2.6.1
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
